Updated AppReducer dispatch with updateAPI to be used if API call for update is needed.

+++ nodestudiojs/src/state/AppReducers.js
@@ -19,23 +19,25 @@ export const ActionTypes = {
 export const AppReducers = (state, action) => {
     let nodes, links;
 
-    switch(action.type) {
-        // Node actions
-        case ActionTypes.ADD_NODE:
-            APIDataService.addNode(Node.export(action.node));
-            break;
-        case ActionTypes.UPDATE_NODE:
-            APIDataService.updateNode(Node.export(action.node));
-            break;
-        case ActionTypes.DELETE_NODE:
-            APIDataService.deleteNode(action.nodeID);
-            break;
-
-        // Link actions
-        case ActionTypes.DELETE_LINK:
-            APIDataService.deleteLink(action.linkID);
-            break;
-        default:
+    if(action.updateAPI === true) {
+        switch(action.type) {
+            // Node actions
+            case ActionTypes.ADD_NODE:
+                APIDataService.addNode(Node.export(action.node));
+                break;
+            case ActionTypes.UPDATE_NODE:
+                APIDataService.updateNode(Node.export(action.node));
+                break;
+            case ActionTypes.DELETE_NODE:
+                APIDataService.deleteNode(action.nodeID);
+                break;
+    
+            // Link actions
+            case ActionTypes.DELETE_LINK:
+                APIDataService.deleteLink(action.linkID);
+                break;
+            default:
+        }
     }
 
     switch(action.type) {
